+++ account_aged_balance_from_partner/partner_view.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  Copyright (C) 2015-2018 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+
+<odoo>
+
+
+<record id="view_partner_form" model="ir.ui.view">
+    <field name="name">account.balance.button.partner.form</field>
+    <field name="model">res.partner</field>
+    <field name="inherit_id" ref="account.partner_view_buttons"/>
+    <field name="arch" type="xml">
+        <button name="open_partner_history" type="object" position="after">
+            <button class="oe_stat_button" type="object"
+                name="open_aged_open_invoices_report"
+                attrs="{'invisible': [('parent_id', '!=', False)]}"
+                icon="fa-money">
+                <div class="o_form_field o_stat_info">
+                    <span class="o_stat_value"><field name="balance"/></span>
+                    <span class="o_stat_text">Account Balance</span>
+                </div>
+            </button>
+        </button>
+    </field>
+</record>
+
+
+</odoo>
