--- partner_aged_open_invoices/__init__.py
@@ -1,3 +1,0 @@
-# -*- encoding: utf-8 -*-
-
-from . import partner
