--- partner_aged_open_invoices/partner_view.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    Copyright (C) 2015 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
--->
-
-<openerp>
-<data>
-
-<record id="view_partner_form" model="ir.ui.view">
-    <field name="name">account.balance.button.partner.form</field>
-    <field name="model">res.partner</field>
-    <field name="inherit_id" ref="base.view_partner_form"/>
-    <!-- Put the button after the "Journal Items" without depending
-    on the view account.partner_view_button_journal_item_count
-    which is linked to the group account.group_account_user -->
-    <field name="priority" eval="22"/>
-    <field name="arch" type="xml">
-        <xpath expr="//div[@name='buttons']" position="inside">
-            <button class="oe_stat_button" type="object"
-                name="open_aged_open_invoices_report"
-                attrs="{'invisible': [('parent_id', '!=', False)]}"
-                icon="fa-money"
-                >
-                <div>
-                    <strong>
-                        <field string="Account Balance" name="balance"
-                            widget="monetary"/>
-                    </strong>
-                    <br/>
-                    Account Balance
-                </div>
-            </button>
-        </xpath>
-    </field>
-</record>
-
-</data>
-</openerp>
